build: fix build.sh for proper CROSS_COMPILE exportation
'export' only passes variables to its children processes, not its parent processes. so having
build.sh call a script to prompt for the toolchain prefix was entirely pointless, since the
build would then be the parent of the script exporting CROSS_COMPILE, and would not get to
see it! instead, export CROSS_COMPILE directly in build.sh, and have anyone else who would like
to build change the definition as they so need.